Redo Component Invokation
If we initialize everything from the start, this causes issues as we only want to render a component as we need to. Some things may not be loaded in yet, as well as the user may not be authenticated - these may be things components rely on, so we want to render them after authentication has taken place.